+++ lib/src/di.dart
@@ -1,28 +1,43 @@
+import 'resource_callback.dart';
+
+/// Dependency injection
+///
+/// Simple and easy dependency injection
+/// for your Dart applications.
+///
+/// Part of Utopia Dart ecosystem,
+/// but can be used independently.
 class DI {
+  static const defaultContext = 'utopia';
   static DI? _instance;
   final Map<String, Map<String, dynamic>> _resources = {};
-  static final Map<String, Map<String, _ResourceCallback>> _resourceCallbacks =
+  static final Map<String, Map<String, ResourceCallback>> _resourceCallbacks =
       {};
 
+  /// Get singleton instance
   static DI get instance {
     _instance ??= DI();
     return _instance!;
   }
 
+  /// Get singleton instance
   static DI get i => DI.instance;
 
+  /// Set a resource callback
   void set(
     String name,
     Function callback, {
     List<String> injections = const [],
-    String context = 'utopia',
+    String context = defaultContext,
   }) {
     _resourceCallbacks[context] ??= {};
     _resourceCallbacks[context]![name] =
-        _ResourceCallback(name, injections, callback, reset: true);
+        ResourceCallback(name, injections, callback, reset: true);
   }
 
-  Map<String, dynamic> getAll(List<String> names, {String context = 'utopia'}) {
+  /// Get all the resources set in the context
+  Map<String, dynamic> getAll(List<String> names,
+      {String context = defaultContext}) {
     final resources = <String, dynamic>{};
     for (final name in names) {
       resources[name] = get(name, context: context);
@@ -30,9 +45,12 @@ class DI {
     return resources;
   }
 
+  /// Get a resource
   dynamic g<T>(String name, {bool fresh = false}) => get<T>(name, fresh: fresh);
 
-  dynamic get<T>(String name, {String context = 'utopia', bool fresh = false}) {
+  /// Get a resource
+  dynamic get<T>(String name,
+      {String context = defaultContext, bool fresh = false}) {
     _resources[context] ??= <String, dynamic>{};
     _resourceCallbacks[context] ??= {};
 
@@ -40,7 +58,7 @@ class DI {
     var resourceCallbacks = _resourceCallbacks[context]!;
     if (resourceCallbacks.isEmpty || resourceCallbacks[name] == null) {
       // use default context when not found in the context
-      resourceCallbacks = _resourceCallbacks['utopia'] ?? {};
+      resourceCallbacks = _resourceCallbacks[defaultContext] ?? {};
     }
     if (resources[name] == null ||
         fresh ||
@@ -63,29 +81,17 @@ class DI {
     throw Exception('Resource type doesn\'t match');
   }
 
-  void reset() {
+  /// Reset cached resources
+  void resetResources([String? context]) {
+    if (context != null) {
+      (_resources[context] ?? {}).clear();
+      return;
+    }
     _resources.clear();
   }
 
-  void resetResources() {
+  /// Resets all the dependencies
+  void reset() {
     _resourceCallbacks.clear();
-  }
-}
-
-class _ResourceCallback {
-  final String name;
-  final List<String> injections;
-  final Function callback;
-  final bool reset;
-
-  _ResourceCallback(
-    this.name,
-    this.injections,
-    this.callback, {
-    this.reset = false,
-  });
-
-  _ResourceCallback copyWith({bool? reset}) {
-    return _ResourceCallback(name, injections, callback, reset: reset ?? false);
   }
 }