+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0
+
+- Improve library namings
+- Fixes for reset methods
+- Add doc comments
+- Support context while resetting cached resources
+
 ## 0.1.0
 
 - Support optional context while setting and getting dependencies